feat: enable package MTD
Only a small subset of the package related to JFFS2 is enabled.
Package LZO is also enabled since it is a dependency.

Package python-docutils has been enabled. This package was manually
disabled in a previous clean-up release. However, said package is
automatically re-enabled after every buildroot run. Let's just keep
it as is.