feat: rework BR2_EXTERNAL structure
There are now two external trees:
  * sbt-open-source
  * sbt-proprietary

The names are self-explanatory. This way, it is easier to distinguish
between the licensing. Furthermore, it is now possible to build an
open-source version of sbtOS (without the proprietary applications).
Most importantly, we can now support multiple board configurations.
Currently, only the Zeus board is supported. The BactoBox board
will be added in a later commit (to the sbt-proprietary tree).