fix: remove deprecated opt.pth file used with Python
Said file used to add /media/system/lib/python3.7/... to Pythons search path
for modules. All the Python applications shipped with sbtOS is now packaged
with the rootfs itself. As such, the opt.pth file is no longer needed.

There shouldn't be any user programs that relies on that path as it was
never documented.