perf: mount USB keys faster by skipping 'disk' device
Note that a 'disk' device (e.g., /dev/sda) can actually contain a file
system. However, it is much more common that a USB key has a partition table.
In this case, the file system will be the partition device (e.g.,
/dev/sda1).

It is possible to support both cases but we have chosen not to do so
for simplicity. See the following for reference:
https://wiki.gentoo.org/wiki/Mdev/Automount_USB/automount