feat: update buildroot to 2019.11.1
We are specifically interested in the following fix:
12121: PyQt5.QtSerialPort and other modules not being built

In our case, this caused PyQt5.QtSvg not to be built. In turn, this
became an annoying known issue to the build process.

With the fix in place, the build process is now without any known issues.

Python is now at version 3.8 (was 3.7).